refactor(solver): split query cache eval impl
